Assignment addition
Now users can fill out a form to add an assignment to the db even tho the quizes are still not an option to be taken on the platform. fixed some while looping mess on the index page. tried to add a function that fetches module data so we dont need sql in the index page but thats still in the works.